Mechanum Test TeleOp
This tests the Mechanum Drive on our testing robot. It includes 3
driving modes - Tank (which we used to use), Mechanum (which moves the
robot sideways and diagonally), and Strafe (which turns while moving
the robot in any direction, moving it in a circle. This is done by
using both joysticks.)